feat: add msw with correct config and fetch specs improved
