--- libs/components/atoms/ButtonStyled/index.ts
@@ -1,1 +1,0 @@
-export * from './ButtonStyled';
