+++ libs/components/atoms/button-styled/index.ts
@@ -1,0 +1,1 @@
+export * from './button-styled';
