+++ libs/components/atoms/index.ts
@@ -1,2 +1,3 @@
-export * from './ButtonStyled';
-export * from './BottomTitleCard';
+export * from './button-styled';
+export * from './bottom-title-card';
+export * from './people-content';
