+++ libs/components/atoms/people-content/index.ts
@@ -1,0 +1,1 @@
+export * from './people-content';
