+++ libs/components/atoms/people-content/people-content.tsx
