+++ libs/mocks/index.ts
@@ -1,2 +1,2 @@
-export * from './swapi';
+export * from './swapi-data';
 export * from './handlers';
