+++ libs/mocks/swapi-data/index.ts
