+++ libs/mocks/swapi-data/people-data.ts
