+++ libs/utils/api-url-handler.ts
@@ -1,0 +1,8 @@
+const prepareUrl = (baseUrl: string, path: string) => {
+  const base = baseUrl || 'http://localhost';
+
+  return base + path;
+};
+
+export const SWAPI_URL = (path: string) =>
+  prepareUrl(process.env['NEXT_PUBLIC_SWAPI_BASE_URL'], path);
