+++ libs/utils/index.ts
@@ -1,2 +1,3 @@
-export * from './capitalizedKeys';
-export * from './getSwapiId';
+export * from './capitalized-keys';
+export * from './get-swapi-id';
+export * from './api-url-handler';
