+++ package.json
@@ -54,9 +54,9 @@
     "@swc/core": "1.2.203",
     "@swc/helpers": "0.4.1",
     "@swc/jest": "0.2.21",
-    "@testing-library/dom": "8.14.0",
+    "@testing-library/dom": "^9.0.0",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
+    "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "14.3.0",
     "@types/jest": "29.0.0",
     "@types/jsonwebtoken": "8.5.9",
